added PUT (#57)
