+++ backend/api/urls.py
@@ -1,8 +1,8 @@
 from django.urls import include, path
 from rest_framework.routers import DefaultRouter
 
-from .views import (AppointmentViewSet, OrganizationViewSet, SpecialtyViewSet,
-                    TownViewSet)
+from .views import (AppointmentViewSet, OrganizationViewSet,
+                    SpecialtyViewSet, TownViewSet)
 
 app_name = 'api'
 router = DefaultRouter()
