+++ backend/api/views/__init__.py
@@ -1,0 +1,4 @@
+from .appointment import AppointmentViewSet
+from .organization import OrganizationViewSet
+from .specialty import SpecialtyViewSet
+from .town import TownViewSet
